# promql: reuse `histogramStatsIterator` where possible, and expose it for other implementations to use