--- a/promql/histogram_stats_iterator.go
+++ b/promql/histogram_stats_iterator.go
@@ -19,7 +19,7 @@ import (
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 )
 
-type histogramStatsIterator struct {
+type HistogramStatsIterator struct {
 	chunkenc.Iterator
 
 	currentH *histogram.Histogram
@@ -34,7 +34,7 @@ type histogramStatsIterator struct {
 // counter reset detection internally and sets the counter reset hint accordingly
 // in each returned histogram objects.
 func NewHistogramStatsIterator(it chunkenc.Iterator) chunkenc.Iterator {
-	return &histogramStatsIterator{
+	return &HistogramStatsIterator{
 		Iterator:  it,
 		currentH:  &histogram.Histogram{},
 		currentFH: &histogram.FloatHistogram{},
@@ -44,7 +44,7 @@ func NewHistogramStatsIterator(it chunkenc.Iterator) chunkenc.Iterator {
 // AtHistogram returns the next timestamp/histogram pair. The counter reset
 // detection is guaranteed to be correct only when the caller does not switch
 // between AtHistogram and AtFloatHistogram calls.
-func (f *histogramStatsIterator) AtHistogram(h *histogram.Histogram) (int64, *histogram.Histogram) {
+func (f *HistogramStatsIterator) AtHistogram(h *histogram.Histogram) (int64, *histogram.Histogram) {
 	var t int64
 	t, f.currentH = f.Iterator.AtHistogram(f.currentH)
 	if value.IsStaleNaN(f.currentH.Sum) {
@@ -76,7 +76,7 @@ func (f *histogramStatsIterator) AtHistogram(h *histogram.Histogram) (int64, *hi
 // AtFloatHistogram returns the next timestamp/float histogram pair. The counter
 // reset detection is guaranteed to be correct only when the caller does not
 // switch between AtHistogram and AtFloatHistogram calls.
-func (f *histogramStatsIterator) AtFloatHistogram(fh *histogram.FloatHistogram) (int64, *histogram.FloatHistogram) {
+func (f *HistogramStatsIterator) AtFloatHistogram(fh *histogram.FloatHistogram) (int64, *histogram.FloatHistogram) {
 	var t int64
 	t, f.currentFH = f.Iterator.AtFloatHistogram(f.currentFH)
 	if value.IsStaleNaN(f.currentFH.Sum) {
@@ -104,7 +104,7 @@ func (f *histogramStatsIterator) AtFloatHistogram(fh *histogram.FloatHistogram) 
 	return t, fh
 }
 
-func (f *histogramStatsIterator) setLastH(h *histogram.Histogram) {
+func (f *HistogramStatsIterator) setLastH(h *histogram.Histogram) {
 	if f.lastH == nil {
 		f.lastH = h.Copy()
 	} else {
@@ -112,7 +112,7 @@ func (f *histogramStatsIterator) setLastH(h *histogram.Histogram) {
 	}
 }
 
-func (f *histogramStatsIterator) setLastFH(fh *histogram.FloatHistogram) {
+func (f *HistogramStatsIterator) setLastFH(fh *histogram.FloatHistogram) {
 	if f.lastFH == nil {
 		f.lastFH = fh.Copy()
 	} else {
@@ -120,7 +120,7 @@ func (f *histogramStatsIterator) setLastFH(fh *histogram.FloatHistogram) {
 	}
 }
 
-func (f *histogramStatsIterator) getFloatResetHint(hint histogram.CounterResetHint) histogram.CounterResetHint {
+func (f *HistogramStatsIterator) getFloatResetHint(hint histogram.CounterResetHint) histogram.CounterResetHint {
 	if hint != histogram.UnknownCounterReset {
 		return hint
 	}
@@ -134,7 +134,7 @@ func (f *histogramStatsIterator) getFloatResetHint(hint histogram.CounterResetHi
 	return histogram.NotCounterReset
 }
 
-func (f *histogramStatsIterator) getResetHint(h *histogram.Histogram) histogram.CounterResetHint {
+func (f *HistogramStatsIterator) getResetHint(h *histogram.Histogram) histogram.CounterResetHint {
 	if h.CounterResetHint != histogram.UnknownCounterReset {
 		return h.CounterResetHint
 	}