This PR builds upon https://github.com/prometheus/prometheus/pull/14097.

It makes two changes:

* It exposes the `histogramStatsIterator` type so that other query engine implementations can use it.
* It adds a `Reset` method that allows the iterator to be reused, reducing allocations and improving performance.

The existing benchmarks show a 2-4% reduction in allocations:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/promql
cpu: Apple M1 Pro
                                                                          │ before.txt  │             after.txt             │
                                                                          │   sec/op    │   sec/op     vs base              │
NativeHistograms/histogram_count_with_short_rate_interval-10                425.7m ± 1%   423.5m ± 1%       ~ (p=0.065 n=6)
NativeHistograms/histogram_count_with_long_rate_interval-10                 721.4m ± 1%   715.2m ± 2%       ~ (p=0.093 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-10   388.5m ± 3%   391.4m ± 2%       ~ (p=0.132 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-10    683.9m ± 1%   680.1m ± 2%       ~ (p=0.093 n=6)

                                                                          │  before.txt   │              after.txt              │
                                                                          │     B/op      │     B/op       vs base              │
NativeHistograms/histogram_count_with_short_rate_interval-10                151.2Mi ±  0%   146.8Mi ±  0%  -2.92% (p=0.002 n=6)
NativeHistograms/histogram_count_with_long_rate_interval-10                 148.0Mi ±  4%   143.5Mi ±  4%  -2.98% (p=0.041 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-10   146.9Mi ± 10%   143.5Mi ± 10%       ~ (p=0.180 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-10    157.8Mi ±  0%   154.4Mi ±  0%  -2.14% (p=0.002 n=6)

                                                                          │ before.txt  │             after.txt             │
                                                                          │  allocs/op  │  allocs/op   vs base              │
NativeHistograms/histogram_count_with_short_rate_interval-10                1.792M ± 0%   1.726M ± 0%  -3.68% (p=0.002 n=6)
NativeHistograms/histogram_count_with_long_rate_interval-10                 1.684M ± 3%   1.618M ± 3%  -3.92% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-10   1.690M ± 2%   1.651M ± 2%  -2.31% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-10    1.611M ± 0%   1.572M ± 0%  -2.42% (p=0.002 n=6)
```

The existing benchmarks don't show a noticeable improvement in latency as the benchmarks run a range query with 240 steps over 3000 series, and so the benchmark time is dominated by other things. 

Modifying the existing benchmarks to run an instant query shows more dramatic results:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/promql
cpu: Apple M1 Pro
                                                                          │ before-instant.txt │         after-instant.txt          │
                                                                          │       sec/op       │   sec/op     vs base               │
NativeHistograms/histogram_count_with_short_rate_interval-10                       23.22m ± 1%   20.74m ± 1%  -10.67% (p=0.002 n=6)
NativeHistograms/histogram_count_with_long_rate_interval-10                        64.70m ± 1%   62.16m ± 1%   -3.94% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-10          15.41m ± 1%   13.53m ± 3%  -12.17% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-10           51.41m ± 1%   49.14m ± 3%   -4.41% (p=0.002 n=6)

                                                                          │ before-instant.txt │          after-instant.txt          │
                                                                          │        B/op        │     B/op      vs base               │
NativeHistograms/histogram_count_with_short_rate_interval-10                      7.538Mi ± 0%   2.112Mi ± 0%  -71.98% (p=0.002 n=6)
NativeHistograms/histogram_count_with_long_rate_interval-10                       9.353Mi ± 0%   3.958Mi ± 1%  -57.68% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-10         6.981Mi ± 0%   2.390Mi ± 0%  -65.77% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-10          8.801Mi ± 0%   4.222Mi ± 1%  -52.02% (p=0.002 n=6)

                                                                          │ before-instant.txt │         after-instant.txt          │
                                                                          │     allocs/op      │  allocs/op   vs base               │
NativeHistograms/histogram_count_with_short_rate_interval-10                      126.25k ± 0%   36.27k ± 0%  -71.27% (p=0.002 n=6)
NativeHistograms/histogram_count_with_long_rate_interval-10                       171.35k ± 0%   81.38k ± 0%  -52.51% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_short_rate_interval-10          99.24k ± 0%   42.26k ± 0%  -57.42% (p=0.002 n=6)
NativeHistogramsCustomBuckets/histogram_count_with_long_rate_interval-10          138.35k ± 0%   81.36k ± 0%  -41.19% (p=0.002 n=6)
```