@beorn7 2025-06-04
@fpetkovski might be interested in this… (and the rest of us might value his verdict ;).